quickshell: Bar hides with borderless fullscreen apps Quickshell does automatically check if a window is fullscreen before hiding it. However, that only works if Hyprland reports the client on the workspace as fullscreen.
In recent versions of Hyprland, they seem to have redone their XWayland rendering in a way where borderless fullscreen applications on XWayland (e.g. plenty of Steam games) will not be considered fullscreen by Hyprland, and hence the `fullscreen` attribute will be reported as `false` if you utilize hyprctl.

So I have updated the activeWindow process in Hyprclients that will check for this issue, and reports as fullscreen if all 3 conditions are met: The window uses xwayland, the window is floating, and the resolution reported matches the XOrg fullscreen resolution (as XOrg doesn't support fractional scaling and reports resolution at a truncated integer scale)